+++ fastapi_nacos_extension/__init__.py
@@ -12,7 +12,7 @@ from .exceptions import (
 )
 from .extension import FastAPINacos
 
-__version__ = "0.1.0"
+__version__ = "0.1.1"
 
 __all__ = [
     "FastAPINacos",